+++ packages/runner/sync-customer-os-data/common/entity/airbyte_raw.go
