+++ packages/runner/sync-customer-os-data/common/entity/sync_status.go
